moves dom-processors into part classes
